ptrtoint: truncate to requested integer size
When passing an int (a machine.Pin) as an interface, and then calling
a method (Set(machine.Pin, bool)) on it, the receiver in the thunk
may be the integer pin value passed as a pointer.
The Set$invoke thunk will ptrtoint the value before calling Set.
Set will icmp the received value against (u8)-1.
Icmp will fail as the raw value of the pointer is of a different size
than the integer value.

To fix this, retrieve the integer size while compiling and pass it as
the second operand.
While interpreting ptrtoint, truncate the pointer value.
AsPointer needs to return the raw value even when it chooses to error
out (due to the value being too small) since in this case we're just
casting bits.